fix to port scanning bug
a bug existed which would cause an exception in port scanning. the
dictionary which kept track of the time a source IP was detected was not
updated when the number of ports accessed is equal to or greater than
the unique ports constant. now, any time this condition is met, the
source dictionary is updated to the current time.